+++ backend/src/main/java/com/mealchemy/vault/dto/VaultFolderRequest.java
@@ -4,12 +4,12 @@ public class VaultFolderRequest {
     
     /* Variable declarations */
 
-    private Long vaultId;
+    private int vaultId;
     private String folderName;
 
     /* Getters */
 
-    public Long getVaultId()
+    public int getVaultId()
     {
         return vaultId;
     }
@@ -21,7 +21,7 @@ public class VaultFolderRequest {
 
     /* Setters */
 
-    public void setVaultId(Long vaultIdIn)
+    public void setVaultId(int vaultIdIn)
     {
         vaultId = vaultIdIn;
     }
